refactor(ui): consolidate AllInModal into shared components directory
Move AllInModal from sessions/[id]/ to components/sessions/ for reuse
across both SessionDetailContent and ActiveSessionContent. Introduce
AllInRecordForModal interface to decouple from tRPC types.
